+++ test/test_helper.rb
@@ -5,6 +5,9 @@ require File.expand_path(File.dirname(__FILE__) + '/../../../test/test_helper')
 require 'minitest/spec'
 require 'minitest/mock'
 require 'minitest/autorun'
+require "minitest/reporters"
+
+Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 DatabaseCleaner.strategy = :transaction
 DatabaseCleaner.clean_with(:truncation)